bool masks

+++ src/levanter/models/lm_model.py
@@ -54,7 +54,7 @@ class LmExample(eqx.Module):
         if ignore_id is not None:
             # we don't compute loss for any tokens matching the ignore index
             ignore_mask = hax.roll(tokens, -1, Pos) != ignore_id
-            loss_mask = loss_mask & ignore_mask
+            loss_mask = loss_mask & ignore_mask.astype(loss_mask.dtype)
 
         loss_mask = loss_mask.astype(jnp.int32)
